fix exception inheritance
